hosts-info/requirements.txt
asyncio
aiohttp
numpy
spidev
pillow
RPi.GPIO
